refactor(bench): separate storage and minibf benchmark subjects
