+++ xtask/archive-bench/MINIBF.md
@@ -1,16 +1,27 @@
 # Minibf performance benchmarks
 
-The archive benchmark tooling has three layers:
+Storage and minibf are separate benchmark subjects, with shared machinery:
 
 | Layer | Entry point | Subject |
 |---|---|---|
 | Storage | `cargo bench --bench archive_backends` | Isolated log scans, local indexes, compressed body reads and scans |
-| Routes and replay | `cargo xtask archive-bench minibf` | Actual minibf routers on fresh Fjall state/archive and a persistent redb WAL |
-| Real-data calibration | `cargo xtask archive-bench minibf-http` | Verified HTTP requests against an independently prepared, stable-tip node |
+| Storage corpus and node runs | `cargo xtask bench storage bench` / `bench storage node` | Compression, I/O and import workloads on matched corpora |
+| Routes and replay | `cargo xtask bench minibf run` | Actual minibf routers on fresh Fjall state/archive and a persistent redb WAL |
+| Real-data calibration | `cargo xtask bench minibf http` | Verified HTTP requests against an independently prepared, stable-tip node |
 
 There is no redb archive comparison or production compression switch. The new
 measurement adapters live in `dolos-testing`; production store and route
 implementations are unchanged.
+
+The Rust modules mirror this split: `xtask::bench::storage` contains storage
+workloads and `xtask::bench::minibf` contains actual minibf router/HTTP workloads.
+Neither subject owns the other. Clocks, counters, provenance, dictionary identity,
+the request load driver and JSONL reporting live under `xtask::bench`; reusable
+chain fixtures and store instrumentation remain in `dolos-testing`.
+`cargo xtask bench report` renders records from either subject or a mixed file.
+The old `archive-bench` command remains a compatibility entry point for existing
+storage scripts, not an umbrella for minibf. Documentation and historical result
+files retain their `xtask/archive-bench/` paths.
 
 ## Build and smoke
 
@@ -19,10 +30,10 @@ Build the xtask from the revision being measured. A globally installed
 
 ```sh
 cargo build --release -p xtask
-target/release/cargo-xtask archive-bench minibf \
+target/release/cargo-xtask bench minibf run \
   --work /path/to/scratch --out smoke.jsonl --run smoke \
   --repeat 1 --requests 10
-target/release/cargo-xtask archive-bench report smoke.jsonl
+target/release/cargo-xtask bench report smoke.jsonl
 ```
 
 Smoke results deliberately lack the repetitions and samples to pass timing gates.
@@ -86,7 +97,7 @@ the cache regime is named `route-primed`, never cold.
 For example, sweep selectivity and pagination while holding the rest fixed:
 
 ```sh
-target/release/cargo-xtask archive-bench minibf \
+target/release/cargo-xtask bench minibf run \
   --work /path/to/scratch --out medium.jsonl --run medium \
   --blocks 256 --log-rows 100000 --pool-stride 32 \
   --page 20 --page-size 5 --requests 1000 --repeat 3 \
@@ -96,7 +107,7 @@ target/release/cargo-xtask archive-bench minibf \
 The storage bench additionally includes logs-only populations and a real Alonzo
 block fixture. Its existing filler-byte cases remain isolated storage guards;
 they are not representative compression-ratio measurements. The existing
-archive-bench `bench` and `node` commands retain their real-corpus and cache
+`bench storage bench` and `bench storage node` commands retain their real-corpus and cache
 experiments.
 
 ## Request load and live replay
@@ -120,7 +131,7 @@ alongside a server. It records completed blocks and whole-roll-forward latency
 (including all lifecycle work), not a fictional index-commit percentile.
 
 ```sh
-target/release/cargo-xtask archive-bench minibf \
+target/release/cargo-xtask bench minibf run \
   --work /path/to/scratch --out live.jsonl --run live \
   --live --blocks 256 --requests 1000 --rates 25,100 \
   --timeout-ms 1000 --write-interval-ms 250 \
@@ -143,12 +154,12 @@ without this harness needs the benchmark-only changes applied to its worktree;
 the runner does not build or modify revisions for you.
 
 ```sh
-target/release/cargo-xtask archive-bench minibf-compare \
+target/release/cargo-xtask bench minibf compare \
   --bin baseline=/path/to/baseline/cargo-xtask \
   --bin candidate=/path/to/candidate/cargo-xtask -- \
   --work /path/to/scratch --out paired.jsonl --run comparison-01 \
   --requests 1000 --repeat 3
-target/release/cargo-xtask archive-bench minibf-check paired.jsonl \
+target/release/cargo-xtask bench minibf check paired.jsonl \
   --max-p95-ratio 1.10 --max-p99-ratio 1.20 --min-throughput-ratio 0.90 \
   --min-repeats 3 --min-samples 1000
 ```
@@ -199,7 +210,7 @@ whole value. Use full responses when practical; a projection verifies only the
 selected fields. Never derive the oracle solely from the candidate.
 
 ```sh
-target/release/cargo-xtask archive-bench minibf-http \
+target/release/cargo-xtask bench minibf http \
   --url http://127.0.0.1:3000 --manifest mainnet.json \
   --server-binary /path/to/baseline/dolos --server-revision <commit> \
   --run mainnet-01 --label baseline --out mainnet.jsonl \
