+++ xtask/src/bench/dictionary.rs
