+++ xtask/src/bench/storage/mod.rs
@@ -1,6 +1,6 @@
 //! Developer benchmarks for the archive's compressed block segments.
 //!
-//! `cargo xtask archive-bench` measures the archive's write and read paths
+//! `cargo xtask bench storage` measures the archive's write and read paths
 //! on real block corpora and renders paired comparisons from the records.
 //! The store-level workloads run the production `dolos_flatfiles` store
 //! beside a modelled sink (one frame per block, or raw bodies) so codec and
@@ -12,14 +12,12 @@
 
 pub mod codec;
 pub mod corpus;
-pub mod dictionary;
-pub mod measure;
-pub mod minibf;
 pub mod node;
 pub mod presets;
-pub mod report;
 pub mod train;
 pub mod workloads;
+
+use super::{command_line, dictionary, measure, parse_list, report};
 
 use std::io::Write;
 use std::path::PathBuf;
@@ -166,23 +164,6 @@ pub enum Cmd {
     /// Drive a dolos binary through import and API workloads
     Node(Box<node::NodeArgs>),
 
-    /// Measure verified minibf routes on fresh persistent fixtures
-    Minibf(Box<minibf::Args>),
-
-    /// Pair benchmark-enabled revisions in alternating order
-    MinibfCompare(Box<minibf::CompareArgs>),
-
-    /// Calibrate verified HTTP workloads against a prepared, stable-tip node
-    MinibfHttp(Box<minibf::http::Args>),
-
-    /// Enforce paired minibf budgets, failing on missing or invalid evidence
-    MinibfCheck {
-        #[arg(required = true)]
-        files: Vec<PathBuf>,
-        #[command(flatten)]
-        budgets: minibf::report::Budgets,
-    },
-
     /// Train a dictionary on a seeded sample of segments
     Train {
         /// directory of raw `NNNNNN.segment` files
@@ -237,17 +218,6 @@ pub enum Cmd {
     Report { files: Vec<PathBuf> },
 }
 
-pub fn parse_list<T: std::str::FromStr>(s: &str) -> anyhow::Result<Vec<T>>
-where
-    T::Err: std::fmt::Display,
-{
-    s.split(',')
-        .map(str::trim)
-        .filter(|p| !p.is_empty())
-        .map(|p| p.parse::<T>().map_err(|e| anyhow::anyhow!("{p}: {e}")))
-        .collect()
-}
-
 fn codecs(spec: &str, dictionary: &str) -> anyhow::Result<Vec<(String, Codec)>> {
     let mut out = Vec::new();
     for name in spec.split(',').map(str::trim).filter(|p| !p.is_empty()) {
@@ -265,24 +235,6 @@ fn codecs(spec: &str, dictionary: &str) -> anyhow::Result<Vec<(String, Codec)>> 
         out.push((name.to_string(), codec));
     }
     Ok(out)
-}
-
-/// The command line as a shell would need it typed, so a recorded command
-/// replays.
-pub fn command_line() -> String {
-    std::env::args()
-        .map(|arg| shell_word(&arg))
-        .collect::<Vec<_>>()
-        .join(" ")
-}
-
-fn shell_word(arg: &str) -> String {
-    let plain = |c: char| c.is_ascii_alphanumeric() || "-_./=,:@+%".contains(c);
-    if !arg.is_empty() && arg.chars().all(plain) {
-        arg.to_string()
-    } else {
-        format!("'{}'", arg.replace('\'', "'\\''"))
-    }
 }
 
 pub fn run(cmd: Cmd) -> anyhow::Result<()> {
@@ -343,16 +295,6 @@ pub fn run(cmd: Cmd) -> anyhow::Result<()> {
             eprintln!("results appended to {}", out.display());
         }
         Cmd::Node(args) => node::run(*args)?,
-        Cmd::Minibf(args) => minibf::run(*args)?,
-        Cmd::MinibfCompare(args) => minibf::compare(*args)?,
-        Cmd::MinibfHttp(args) => minibf::http::run(*args)?,
-        Cmd::MinibfCheck { files, budgets } => {
-            budgets.validate()?;
-            let records = report::load(&files)?;
-            let (rendered, passed) = minibf::report::assess(&records, &budgets);
-            print!("{rendered}");
-            anyhow::ensure!(passed, "minibf comparison did not pass; see report");
-        }
         Cmd::Train {
             corpus,
             segments,
