+++ xtask/src/lib.rs
@@ -1,3 +1,3 @@
 //! Library half of the xtask binary: the pieces its tests drive directly.
 
-pub mod archive_bench;
+pub mod bench;
